bpf: Add an additional optional netns stage to BPF tests
NETNS(<progtype>, <name>) adds the ability to configure a network namespace.
NETNS is executed first and uses sandbox_netns_map to pass the namespace name to user space.
A network namespace with the provided name is then created and removed once the test finishes.

The primary motivation for this feature is to prevent unintended side effects
that may occur when tests are executed in the host namespace.

Signed-off-by: Jack-R-lantern <tjdfkr2421@gmail.com>