bpf: migrate runtime config (enable_host_firewall)
enable_host_firewall has been migrated to a runtime configuration.

Previously, the code was compiled only when the option was enabled,
but after this change it is always compiled.

As a result, some functions and variables may become unused,
so __maybe_unused has been added to those parts as a temporary measure.

Signed-off-by: Jack-R-lantern <tjdfkr2421@gmail.com>